fix: Discount Id int -> long, 주문내역조회 오류 수정

+++ src/main/java/com/smplatform/product_service/domain/discount/dto/DiscountResponseDto.java
@@ -14,7 +14,7 @@ public class DiscountResponseDto {
     @AllArgsConstructor
     public static class DiscountInfo {
 
-        private int discountId;
+        private long discountId;
 
         private String discountName;
